Use RemoteMediator to get Post data from database and api

+++ NewsFeedApp/data/src/main/java/com/example/data/remote/api/PostService.kt
@@ -21,7 +21,7 @@ interface PostService {
     @GET("posts")
     suspend fun getPosts(
         @Query("limit") limit: Int = 10,
-        @Query("cursor") cursor: Int? = null,
+        @Query("cursor") cursor: Long? = null,
     ): PostListApiResponse
 
     @POST("posts/interaction")
