+++ NewsFeedApp/domain/src/main/java/com/example/domain/model/PaginatedPosts.kt
@@ -6,6 +6,6 @@ data class PaginatedPosts(
 )
 
 data class PaginationMetaData(
-    val nextCursor: Int?,
+    val nextCursor: Long?,
     val hasMore: Boolean
 )
